// src/modules/items/application/types/ItemCategory.ts
export enum ItemCategory {
  AMPLITUNER = "amplituner",
  INTERFACE = "interface",
  MICROPHONE = "microphone",
  MIXER = "mixer",
  SOUND_RECORDER = "sound_recorder",
  SPEAKER = "speaker",
  HEADPHONES = "headphones",
  DIGITAL_SOUND_PROCESSOR = "digital_sound_processor",
  PREAMPLIFIER = "preamplifier",
  CAMERA = "camera",
  VIDEO_CAMERA = "video_camera",
  CONVERTER = "converter",
  CAMERA_LENS = "camera_lens",
  FILTER = "filter",
  LIGHTING = "lighting",
  PLAYER = "player",
  SCREEN = "screen",
  TRIPOD = "tripod",
  PROJECTOR = "projector",
  GLASSES_3D = "glasses_3d",
  OTHER = "other",
}